moved global to framework/global
