+++ CMakeLists.txt
@@ -787,7 +787,6 @@ if (BUILD_TELEMETRY_MODULE)
       add_subdirectory(thirdparty/google_analytics)
 endif (BUILD_TELEMETRY_MODULE)
 
-add_subdirectory(global)
 add_subdirectory(importexport)
 add_subdirectory(mscore)
 add_subdirectory(main)
