+++ framework/global/modularity/modularity.cmake
@@ -1,4 +1,6 @@
 
 set (MODULARITY_SRC
     ${CMAKE_CURRENT_LIST_DIR}/imodulesetup.h
+    ${CMAKE_CURRENT_LIST_DIR}/serviceinjector.h
+    ${CMAKE_CURRENT_LIST_DIR}/servicesresolver.h
 )
