+++ framework/global/modularity/servicesresolver.h
