+++ main/CMakeLists.txt
@@ -43,7 +43,7 @@ else (APPLE)
 endif (APPLE)
 
 include_directories(
-      ${PROJECT_SOURCE_DIR}/global
+      ${PROJECT_SOURCE_DIR}/framework/global
       ${PROJECT_BINARY_DIR}
       ${QTSINGLEAPPLICATION_INCLUDE_DIRS}
       )
